requirements.txt
torch
torchvision
numpy
toolz
pytorch-ignite
Pillow>5.0.0
xmltodict
googledrivedownloader
pybind11
sklearn
pytorchcv
yacs
thop
timm
hhutil